Remove redundant client_extension.go
There's no need to have a client extension file, we can put all
things to one client file.
